+++ libhmsbeagle/CPU/BeagleCPUAVXImpl.hpp
@@ -567,8 +567,8 @@ BeagleImpl* BeagleCPUAVXImplFactory<BEAGLE_CPU_FACTORY_GENERIC>::createImpl(int 
                                              int scaleBufferCount,
                                              int resourceNumber,
                                              int pluginResourceNumber,                                             
-                                             long preferenceFlags,
-                                             long requirementFlags,
+                                             long long preferenceFlags,
+                                             long long requirementFlags,
                                              int* errorCode) {
 
     if (!CPUSupportsAVX())
