+++ libhmsbeagle/CPU/BeagleCPUActionImpl.h
@@ -166,8 +166,8 @@ namespace beagle {
                                        int scaleBufferCount,
                                        int resourceNumber,
                                        int pluginResourceNumber,
-                                       long preferenceFlags,
-                                       long requirementFlags);
+                                       long long preferenceFlags,
+                                       long long requirementFlags);
 
             virtual int setPartials(int bufferIndex,
                             const double* inPartials);
@@ -275,8 +275,8 @@ public:
                                    int scaleBufferCount,
                                    int resourceNumber,
                                    int pluginResourceNumber,
-                                   long preferenceFlags,
-                                   long requirementFlags,
+                                   long long preferenceFlags,
+                                   long long requirementFlags,
                                    int* errorCode);
 
     virtual const char* getName();
